docs(grafana_dashboard): update for anonymous access and formatting
Update the documentation to reflect the necessary configurations in the `grafana.ini`
file for enabling embedding and anonymous access. Also, clarify the prerequisites for
using Grafana with HertzBeat and VictoriaMetrics. The changes include specifying the
`allow_embedding` and `[auth.anonymous]` settings and correcting the formatting of the`VictoriaMetrics` database reference.